Added local python script for sending playback via api

+++ local_script/run.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+DEBUG=true
 
 report_error(){
     notify-send "WhatAmIListeningTo" "ERROR: $1"
@@ -12,5 +13,11 @@ fi
 
 source "$DIR/.venv/bin/activate"
 
-python "$DIR/run.py" & spotify
-pkill -f "$DIR/run.py"
+if [ "$DEBUG" = true ] ; then
+    export API_URL="localhost:8000/"
+    python "$DIR/run.py"
+else
+#    export API_URL="will be done soon"
+    python "$DIR/run.py" & spotify
+    pkill -f "$DIR/run.py"
+fi
